Fix to reduce reindex memory requirements.
The `reindex()` method loaded the entire collection being reindexed,
which caused an out-of-memory condition on large collections. Refactored
it to load the collection in chunks. (It was already doing an
ElasticSearch bulk update in chunks, so just loaded the records required
for each update chunk.)

Removed the optional `start` parameter from `reindex()`. Belief is that
it was added for use in development, and was never used in production
code.

Also did some cosmetic cleanup: Fixed some things ESLint was complaining
about. Used standard Promise instead of Bluebird.