Retry on HTTP 502 errors.
Added 502 errors to retry logic (in addition to 429 errors).

Changed default retry factor from 2 to 4; this throttles the retries
more aggressively, and extends the total retry duration. Total duration
is now 17 sec by default; hopefully long enough to accommodate a restart
of the ElasticSearch service.

Added `retryFactor` as a configurable setting.

Tweaked the logging to report which operation was being retried.